feat: landingPage 파일명 SplashScreen으로 변경 및 이동, 디자인 및 렌더링 수정
